CA-814 - homepage - time to act - fixed the FAQ link.
The built in Link component of react-router-dom does not support
hash navigations, so it does not scroll. Since the application
already uses react-router-hash-link, which exposes HashLink,
an in place scroll-to-hash-supporting replacement of Link,
GhostButton was changed to use HashLink.
This makes the Read More link in the Time To Act section in the
homepage navigate to the about page and scroll to the FAQ section.